[refactor] move countryTrends to twitterAPI folder
